+++ packages/expo-crypto/src/Crypto.ts
@@ -7,6 +7,11 @@ declare const global: any;
 
 export * from './Crypto.types';
 
+/**
+ * @hidden
+ */
+export * as AES from './aes';
+
 class CryptoError extends TypeError {
   code = 'ERR_CRYPTO';